Task2 - AZURE CUSTOM VISION Image Classification/Azure Custom vision   IMAGE_CLASSIFIER (task-2)/requirement.txt
azure-cognitiveservices-vision-customvision
msrest==0.6.21
matplotlib
Flask==2.0.1